==> notebooks/1_ajs_addtractstonypd.ipynb <==
# libraries
import pickle
import pandas as pd
import geopandas as gpd

Goal: Place each Level 3 stop in a census tract area

# Load nypd_geo
input_filepath = './data/interim/nypd_geo.pkl'

# Open the pickle file in read mode
with open(input_filepath, 'rb') as file:
    nypd_geo = pickle.load(file)


# load census geodata
census_tracts = gpd.read_file('./data/processed/geo/censustracts2020.geojson') 

# Reproject census to EPSG:2263
census_tracts = census_tracts.to_crs(epsg=2263)
print(census_tracts.crs)

# Make a list of census tract column names
column_names = census_tracts.columns.tolist()

# Reset index of both GeoDataFrames
nypd_geo = nypd_geo.reset_index(drop=True)
census_tracts = census_tracts.reset_index(drop=True)

# join the geo dataframes
nypd_geo =nypd_geo.sjoin_nearest(census_tracts)

# save to csv
nypd_geo.to_csv('./data/processed/geo/nypd_geo.csv', index=False)

# Define the file path
output_filepath = './data/processed/geo/nypd_geo.pkl'

# Open the file in write mode
with open(output_filepath, 'wb') as file:
    # Save the processed DataFrame as a pickle object
    pickle.dump(nypd_geo, file)

# remove 'BoroCT2020' and 'BoroName' from list
column_names.remove('BoroCT2020')
column_names.remove('BoroName')

# remove columns in column_names from nypd_geo
nypd_tracts = nypd_geo.drop(column_names, axis=1)

# save to csv
nypd_tracts.to_csv('./data/processed/numeric/nypd.csv', index=False)

# Define the file path
output_filepath = './data/processed/numeric/nypd.pkl'

# Open the file in write mode
with open(output_filepath, 'wb') as file:
    # Save the processed DataFrame as a pickle object
    pickle.dump(nypd_tracts, file)

# Summarize the nypd data by census tract

#summarize by race
nypd_summary = nypd_tracts.groupby('BoroCT2020')['SUSPECT_RACE_DESCRIPTION'].value_counts().unstack().fillna(0)

nypd_summary = nypd_summary.reset_index()

# save to csv
nypd_summary.to_csv('./data/processed/numeric/nypd_byrace.csv', index=False)

# Define the file path
output_filepath = './data/processed/numeric/nypd_byrace.pkl'

# Open the file in write mode
with open(output_filepath, 'wb') as file:
    # Save the processed DataFrame as a pickle object
    pickle.dump(nypd_summary, file)
